- changed initialization to a "modules subscribe themselves" pattern
