+++ defrag/__main__.py
@@ -14,14 +14,10 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-from defrag.modules.db.redis import RedisPool
 import uvicorn
 import importlib
 from defrag import app, LOGGER, pretty_log
 from defrag.modules import ALL_MODULES
-from defrag.modules.helpers import Query
-from defrag.modules.helpers.caching import CacheStrategy, RedisCacheStrategy
-from defrag.modules.helpers.services_manager import Run, ServiceTemplate, ServicesManager
 
 IMPORTED = {}
 
@@ -39,32 +35,6 @@ def main() -> None:
             # NO_TWO_MODULES
             raise Exception(
                 "Can't have two modules with the same name! Please change one")
-
-
-@app.on_event("startup")
-async def startup() -> None:
-    with RedisPool() as conn:
-        conn.flushall()
-    reddit_strategy = CacheStrategy(RedisCacheStrategy(
-        "reddit_default", True, True, 300, None, None), None)
-    twitter_strategy = CacheStrategy(RedisCacheStrategy(
-        "twitter_default", True, True, 300, None, None), None)
-    reddit = ServiceTemplate("reddit", reddit_strategy, None, None, None, None)
-    twitter = ServiceTemplate(
-        "twitter", twitter_strategy, None, None, None, None)
-    ServicesManager.initialize(twitter, reddit)
-    pretty_log("Finished startup with services",
-               str(ServicesManager.services.list_all()))
-
-
-@app.get("/twitter")
-async def get_twitter():
-    return await Run.query(Query(service="twitter"))
-
-
-@app.get("/reddit")
-async def get_reddit():
-    return await Run.query(Query(service="reddit"))
 
 
 if __name__ == "__main__":
