--- defrag/tests/unittests/test_news_reddit_twitter.py
@@ -1,1 +1,0 @@
-#deprecated, will rewrite
